when direction of render is changed, add the margin to the icons (#26)
When the browser direction is Right to Left or Left to Right we need to override the margin-right which is present

Issue: When the user of language Hebrew or arabic is viewed the direction of render would be "rtl" and the margins on the icons were making the render on "rtl" wonky"

Fix: To use -webkit-margin-start/ -webkit-margin-end which will , If the writing direction is left-to-right, this value overrides margin-left. If the writing direction is right-to-left, this value overrides margin-right.